Insane refactoring
